+++ util/conf.js
@@ -23,7 +23,7 @@ function parse(conf){
 function _hostname(){
     let hostname = env.CONFIG_HOSTNAME || os.hostname();
     return hostname.replace(/\.hola\.org$/, '').replace(/\.localdomain$/, '')
-        .replace(/\.home$/, '');
+        .replace(/\.local$/, '').replace(/\.home$/, '');
 }
 
 function init(){